moved the keyToLabel and UUID function out of helpers and into their respective files. due to vercel build failing and error log.
