+++ src/routes/(private)/cms/units/[unit_id]/[category]/[subcategory]/[option]/AddPhotoDropdown.svelte
@@ -11,7 +11,6 @@
     setDoc,
   } from "firebase/firestore";
   import { createEventDispatcher } from "svelte";
-  import { newUUID } from "$lib/helpers";
   import type { Unit } from "$lib/types";
   import { slide } from "svelte/transition";
 
@@ -134,6 +133,17 @@
     }[]
   };
     */
+  }
+
+  function newUUID(): string {
+    // Alphanumeric characters
+    const chars =
+      "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
+    let autoId = "";
+    for (let i = 0; i < 20; i++) {
+      autoId += chars.charAt(Math.floor(Math.random() * chars.length));
+    }
+    return autoId;
   }
 </script>
